+++ friends_server/src/main/kotlin/com/friends/chat/websocket/ChatWebSocketHandler.kt
@@ -9,14 +9,20 @@ import org.springframework.web.socket.CloseStatus
 import org.springframework.web.socket.TextMessage
 import org.springframework.web.socket.WebSocketSession
 import org.springframework.web.socket.handler.TextWebSocketHandler
+import java.util.concurrent.ConcurrentHashMap
+import java.util.concurrent.CopyOnWriteArraySet
 
 @Component
 class ChatWebSocketHandler(
     private val messageService: MessageService,
 ) : TextWebSocketHandler() {
+    // 채팅방 ID를 키로 하고, 각 채팅방의 세션을 Set으로 저장
+    private val chatRooms: MutableMap<Long, MutableSet<WebSocketSession>> = ConcurrentHashMap()
+
     override fun afterConnectionEstablished(session: WebSocketSession) {
         val chatRoomId = getChatRoomId(session)
-        messageService.connectChatRoom(chatRoomId, session)
+        chatRooms.computeIfAbsent(chatRoomId) { CopyOnWriteArraySet() }
+        chatRooms[chatRoomId]?.add(session)
         // TODO : 채팅방 입장 실패 시 에러 처리
     }
 
@@ -31,7 +37,13 @@ class ChatWebSocketHandler(
          * 채팅 웹소켓을 통해 보내는 메세지는 TEXT 타입만 있다고 가정합니다.
          * 이미지의 경우 웹소켓이 아닌 REST API 를 톹ㅇ해 이미지를 업로드하고 이미지 URL 을 채팅방에 보내는 방식으로 구현합니다. // TODO : 채팅방 내에서 이미지 전송하는 API 구현
          */
-        messageService.sendMessage(chatRoomId, memberId, chatMessage.message, MessageType.TEXT)
+        val sendMessageDto = messageService.saveMessage(chatRoomId, memberId, chatMessage.message, MessageType.TEXT)
+        val sessions = chatRooms[chatRoomId]
+        sessions?.forEach { session ->
+            if (session.isOpen) {
+                session.sendMessage(TextMessage(JsonUtil.toJson(sendMessageDto)))
+            }
+        }
         // TODO : 메세지 전송 실패 시 에러 처리
     }
 
@@ -41,6 +53,7 @@ class ChatWebSocketHandler(
     ) {
         val chatRoomId = getChatRoomId(session)
         val memberId = getMemberId(session)
+        chatRooms[chatRoomId]?.remove(session) // 세션 제거
         messageService.disconnectChatRoom(chatRoomId, memberId, session)
         // TODO : 채팅방 나가기 실패 시 에러 처리
     }
